JUZU-48 : Remove unnecessary bottleneck when retrieving Requests
Actually, there a big bottleneck JUZU applications
Each time a JUZU application is displayed in a page, all other HTTP requests that wants to display any other page that uses JUZU will be blocked because of one operation in JUZU application lifecycle.
In fact, inside the "retrieve the response" phase, the operation "get the filters beans associated to the Stage type(Unmarshaling, Handler, Lifecycle, Invoke)" is synchronized (blocker, one operation at a time).
The filters are SINGLETON beans, so we should put those beans in the JUZU controller context once for all.